old_deprecated_approach/requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning & Data Science
scikit-learn>=1.0.0
statsmodels>=0.12.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Dashboard
streamlit>=1.10.0

# LangChain and AI
langchain>=0.0.200
langchain-openai>=0.0.1
langchain-community>=0.0.1
openai>=0.27.0
langgraph>=0.0.50
langchain-anthropic>=0.1.0
anthropic>=0.20.0

# Development & Testing
jupyter>=1.0.0
ipykernel>=6.0.0
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.10.0
typing-extensions>=4.0.0
PyYAML>=6.0

# Optional: For advanced optimization
cvxpy>=1.2.0
pulp>=2.6.0

# Docker support (future)
# gunicorn>=20.1.0

drawdata
lets_plot
langchain-anthropic
